+++ CHANGELOG.md
@@ -3,6 +3,7 @@
   * New built-in status effects for single-use boost/setback dice
   * End-to-end tests for new functionality are in place. Hopefully, this will reduce bugs going forward ([#1837](https://github.com/StarWarsFoundryVTT/StarWarsFFG/issues/1837))
   * Add Force Power Roll button ([#1827](http://overlord.wrycu.com:12121/game))
+  * Added support for universal specializations (requires a re-import of the data) ([#1778](https://github.com/StarWarsFoundryVTT/StarWarsFFG/issues/1778))
 * Fixes:
   * Talent modifiers on specializations are now properly cleaned up when a new talent overrides an old one
   * Force powers, signature abilities, and specializations now use a unified UI to editing upgrades/talents on their sheets ([#1828](https://github.com/StarWarsFoundryVTT/StarWarsFFG/issues/1828))